+++ exchange-simulator/Models/OrderProcessingResult.cs
@@ -2,7 +2,7 @@ namespace exchange_simulator.Models;
 
 public sealed record OrderProcessingResult(
     IReadOnlyList<Trade> Trades,
-    decimal RemainingSize,
+    long RemainingSize,
     bool IsResting
     
     );
